refactor(bpf): extract hook-agnostic datapath helpers (Phase T, T1)
Behavior-neutral preparation for the tc-ingress datapath variant: the
packet-level helpers both hook types need move to a new bpf/src/
datapath.rs, parameterized on (start, end) packet-bounds scalars
instead of XdpContext — TTL/hop-limit decrement + RFC 1624 checksum
patch, l4_ports, icmpv6_type, the SYN-flag test, the full mss-clamp
option walk, bytes_to_u32x4, and the match-subset counter bookkeeping.
fib.rs likewise swaps its XdpContext parameter for the same scalar
bounds, making the whole custom-FIB lookup path hook-agnostic.

Deliberately no Ctx trait and no generics: monomorphized abstraction
would likely verify, but this codebase's verifier history (the inline
mandate, the Option<u16> spill, the memset-libcall exclusions) argues
for the dumbest possible sharing — scalars in, scalars out, everything
still #[inline(always)]. The emitted programs are unchanged: callers
pass ctx.data()/ctx.data_end() at the same points the callees
previously read them, and the load-bearing verifier patterns
(pkt-minus-pkt re-derivation in mss_clamp_tcp, bounded option walk,
bounds-check shapes) move verbatim.

XDP-only pieces stay put: ptr_at/ptr_mut_at (XdpContext accessors),
the VLAN byte-shift choreography (bpf_xdp_adjust_head), the two-stage
tail-call plumbing, and compare_and_bump (bpf_fib_lookup-typed).
